Send keys for a UK keyboard when host runs Windows
I aim to keep my Mac on the 'Unicode Hex Input' source, which is based
on the US keyboard layout. The Windows machines that I use expect a UK
keyboard by default.

Use OS detection to work out whether the host is a Windows machine and,
if so, enable various key overrides to send codes appropriate for a UK
keyboard.