Use macros to simulate shifted UK keys
When a shifted key appears in the keymap, send its UK keyboard
equivalent using an override action based on a macro instead of a simple
replacement. This is to avoid Microsoft Remote Desktop dropping the
shift modifier.

This improves on the previous implementation that was based on Unicode,
which did not work with the OS's auto-repeat, and did not work
everywhere in Windows (presumably a limitation of the HexNumpad input
mode).