Fix keymaps following upstream QMK updates
Fix use of keycodes that have been renamed.

Switch to using the rp2040_ce converter with the 3x5/elitec Charybdis
variant, because the v2/splinky_3 variant has been removed. The
documentation for the breaking change helpfully did not mention which
converters to use, and the documentation for the converters does not
mention Splinky at all. I found which converter to use by searching in
the QMK Discord.

The upstream refactoring of the Charybdis firmware also removed the
MASTER_RIGHT setting, previously defined in v2/splinky_3/config.h. This
was also not mentioned in the documentation for the breaking change.
Unfortunately, with the left-hand side as the primary click-drag does
not work (I don't know why), so I have reinstated MASTER_RIGHT in my
keymap.